requirements.txt
aiosqlite==0.17.0
anyio==3.4.0
asgiref==3.4.1
autopep8==1.6.0
click==8.0.3
databases==0.5.3
fastapi==0.70.1
greenlet==1.1.2
h11==0.12.0
httptools==0.3.0
idna==3.3
ormar==0.10.23
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pydantic==1.8.2
python-dotenv==0.19.2
PyYAML==6.0
sniffio==1.2.0
SQLAlchemy==1.4.28
starlette==0.16.0
toml==0.10.2
typing-extensions==4.0.1
uvicorn==0.16.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
python-decouple==3.5
PyJWT==1.7.1
